+++ test/support/postgres/test_repo.ex
@@ -54,6 +54,11 @@ if Code.ensure_loaded?(Ecto.Adapters.SQL) and Code.ensure_loaded?(Postgrex) do
     use Ecto.Repo, otp_app: :docket, adapter: Ecto.Adapters.Postgres
   end
 
+  defmodule Docket.Postgres.SharedBackendTestRepo do
+    @moduledoc false
+    use Ecto.Repo, otp_app: :docket, adapter: Ecto.Adapters.Postgres
+  end
+
   defmodule Docket.Postgres.BackendSandboxTestRepo do
     @moduledoc false
     use Ecto.Repo, otp_app: :docket, adapter: Ecto.Adapters.Postgres